+++ app/src/main/java/com/app/arcabyolimpo/domain/model/workshops/Workshop.kt
@@ -6,7 +6,6 @@ import com.app.arcabyolimpo.domain.model.supplies.SupplyBatch
  * Domain model representing a supply item in the system.
  *
  * @property id The unique identifier of the workshop.
- * @property idTraining The unique identifier of the training.
  * @property idUser The unique identifier of the user.
  * @property url The url of the image to be upload.
  * @property nameWorkshop The name of the new workshop.
@@ -15,11 +14,11 @@ import com.app.arcabyolimpo.domain.model.supplies.SupplyBatch
  * @property status The status of the workshop, active = 1, inactive = 0.
  * @property description The description of how is the workshop.
  * @property date The day when the workshop is going to take place.
+ * @property videoTraining The video of the training depending on the workshop
  */
 
 data class Workshop(
     val id: String?,
-    val idTraining: String?,
     val idUser: String?,
     val nameWorkshop: String?,
     val url: String?,
@@ -27,5 +26,6 @@ data class Workshop(
     val finishHour: String?,
     val status: Int,
     val description: String?,
-    val date: String?
+    val date: String?,
+    val videoTraining: String?
 )